feat(aliases): organizing aliases and rules

+++ lib/aliases/tmux/tmux.aliases.sh
@@ -5,11 +5,13 @@
 # License: MIT
 
 # 🆃🅼🆄🆇 🅰🅻🅸🅰🆂🅴🆂
-alias tm='tmux'                     # tm: Start tmux.
-alias tma="tmux attach-session"     # tma: Attach to a tmux session.
-alias tmat="tmux attach-session -t" # tmat: Attach to a tmux session with name.
-alias tmks='tmux kill-session -a'   # tmks: Kill all tmux sessions.
-alias tml='tmux list-sessions'      # tml: List tmux sessions.
-alias tmn="tmux new-session"        # tmn: Start a new tmux session.
-alias tmns="tmux new -s"            # tmns: Start a new tmux session with name.
-alias tms='tmux new-session -s'     # tms: Start a new tmux session.
+if command -v tmux &>/dev/null; then
+  alias tm='tmux'                     # tm: Start tmux.
+  alias tma="tmux attach-session"     # tma: Attach to a tmux session.
+  alias tmat="tmux attach-session -t" # tmat: Attach to a tmux session with name.
+  alias tmks='tmux kill-session -a'   # tmks: Kill all tmux sessions.
+  alias tml='tmux list-sessions'      # tml: List tmux sessions.
+  alias tmn="tmux new-session"        # tmn: Start a new tmux session.
+  alias tmns="tmux new -s"            # tmns: Start a new tmux session with name.
+  alias tms='tmux new-session -s'     # tms: Start a new tmux session.
+fi
